refactored code into separate files
