+++ ser/constants.py
@@ -1,0 +1,4 @@
+from pathlib import Path
+
+PROJECT_ROOT = Path(__file__).parent.parent
+DATA_DIR = PROJECT_ROOT / "data"
